Update for mtime 2.0
Note that the fixes in this commit are not ideal: indeed
Mtime.Span.to_float_ns is only accurate up to a span of
about 104 days (due to floats limited precision). A better
fix would be to convert to float as late as possible (and
potentially convert into seconds rather than nanoseconds).